# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "HTTP API for landlord application",
        "title": "Landlord API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/tenants": {
            "get": {
                "description": "Returns a paginated list of tenants",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tenants"
                ],
                "summary": "List all tenants",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Maximum number of results (default 50)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of results to skip (default 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include archived tenants in results",
                        "name": "include_deleted",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of tenants",
                        "schema": {
                            "$ref": "#/definitions/models.ListTenantsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid pagination parameters",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new tenant with provided configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tenants"
                ],
                "summary": "Create a new tenant",
                "parameters": [
                    {
                        "description": "Tenant creation request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateTenantRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Tenant created successfully",
                        "schema": {
                            "$ref": "#/definitions/models.TenantResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request or validation error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Tenant name already exists",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/tenants/{id}": {
            "get": {
                "description": "Retrieves a specific tenant resource",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tenants"
                ],
                "summary": "Get a tenant by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant identifier (UUID or name)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tenant found",
                        "schema": {
                            "$ref": "#/definitions/models.TenantResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid tenant identifier format",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Tenant not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates properties of an existing tenant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tenants"
                ],
                "summary": "Update a tenant",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant identifier (UUID or name)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Tenant update request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateTenantRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tenant updated successfully",
                        "schema": {
                            "$ref": "#/definitions/models.TenantResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request or validation error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Tenant not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a specific tenant resource",
                "tags": [
                    "tenants"
                ],
                "summary": "Delete a tenant",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant identifier (UUID or name)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tenant deleted",
                        "schema": {
                            "$ref": "#/definitions/models.TenantResponse"
                        }
                    },
                    "202": {
                        "description": "Tenant deletion initiated",
                        "schema": {
                            "$ref": "#/definitions/models.TenantResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid tenant identifier format",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Tenant not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/tenants/{id}/archive": {
            "post": {
                "description": "Archives a tenant by removing compute resources and retaining the record",
                "tags": [
                    "tenants"
                ],
                "summary": "Archive a tenant",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant identifier (UUID or name)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tenant already archived",
                        "schema": {
                            "$ref": "#/definitions/models.TenantResponse"
                        }
                    },
                    "202": {
                        "description": "Tenant archival initiated",
                        "schema": {
                            "$ref": "#/definitions/models.TenantResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid tenant identifier format",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Tenant not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Invalid state transition",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/compute/config": {
            "get": {
                "description": "Returns the active compute provider and its compute_config schema (and defaults if available)",
                "produces": [
                    "application/json"
                ],
                "summary": "Get compute config discovery",
                "tags": [
                    "compute"
                ],
                "responses": {
                    "200": {
                        "description": "Compute config discovery",
                        "schema": {
                            "$ref": "#/definitions/models.ComputeConfigDiscoveryResponse"
                        }
                    },
                    "500": {
                        "description": "Compute provider not configured",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.CreateTenantRequest": {
            "type": "object",
            "required": [
                "image",
                "name"
            ],
            "properties": {
                "annotations": {
                    "description": "Annotations are key-value pairs for metadata",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "compute_config": {
                    "description": "ComputeConfig is provider-specific configuration (Docker, ECS, K8s, etc.)\nValidated by the active compute provider",
                    "type": "object",
                    "additionalProperties": true
                },
                "image": {
                    "description": "Image is the container image the tenant should run",
                    "type": "string"
                },
                "labels": {
                    "description": "Labels are key-value pairs for organizing tenants",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "description": "Name is the unique human-friendly name for the tenant",
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "description": "Details provides additional context about the error",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "error": {
                    "description": "Error is the error message",
                    "type": "string"
                },
                "request_id": {
                    "description": "RequestID is the correlation ID from the request context",
                    "type": "string"
                }
            }
        },
        "models.ListTenantsResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "description": "Number of items per page",
                    "type": "integer"
                },
                "offset": {
                    "description": "Starting position",
                    "type": "integer"
                },
                "tenants": {
                    "description": "Tenants is the array of tenant resources",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TenantResponse"
                    }
                },
                "total": {
                    "description": "Pagination metadata",
                    "type": "integer"
                }
            }
        },
        "models.TenantResponse": {
            "type": "object",
            "properties": {
                "annotations": {
                    "description": "Annotations are key-value pairs for metadata",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "compute_config": {
                    "description": "ComputeConfig is the provider-specific compute configuration",
                    "type": "object",
                    "additionalProperties": true
                },
                "created_at": {
                    "description": "CreatedAt is when the tenant was first created",
                    "type": "string"
                },
                "desired_config": {
                    "description": "DesiredConfig is tenant-specific configuration",
                    "type": "object",
                    "additionalProperties": true
                },
                "desired_image": {
                    "description": "DesiredImage is the container image the tenant should run",
                    "type": "string"
                },
                "id": {
                    "description": "ID is the internal database identifier (UUID)",
                    "type": "string"
                },
                "labels": {
                    "description": "Labels are key-value pairs for filtering and grouping",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "description": "Name is the user-facing stable identifier",
                    "type": "string"
                },
                "observed_config": {
                    "description": "ObservedConfig is the actual configuration applied to running resources",
                    "type": "object",
                    "additionalProperties": true
                },
                "observed_image": {
                    "description": "ObservedImage is the container image currently running",
                    "type": "string"
                },
                "observed_resource_ids": {
                    "description": "ObservedResourceIDs contains provider-specific resource identifiers",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "status": {
                    "description": "Status represents where the tenant is in its lifecycle",
                    "type": "string"
                },
                "status_message": {
                    "description": "StatusMessage provides human-readable context about current status",
                    "type": "string"
                },
                "updated_at": {
                    "description": "UpdatedAt is when the tenant was last modified",
                    "type": "string"
                },
                "version": {
                    "description": "Version is the concurrency control version",
                    "type": "integer"
                },
                "workflow_execution_id": {
                    "description": "WorkflowExecutionID is the ID of the current or last workflow execution",
                    "type": "string"
                }
            }
        },
        "models.UpdateTenantRequest": {
            "type": "object",
            "properties": {
                "annotations": {
                    "description": "Annotations are key-value pairs for metadata (optional for updates)",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "compute_config": {
                    "description": "ComputeConfig is provider-specific configuration (optional for updates)\nIf provided, will be validated by the active compute provider",
                    "type": "object",
                    "additionalProperties": true
                },
                "image": {
                    "description": "Image is the container image the tenant should run (optional for updates)",
                    "type": "string"
                },
                "labels": {
                    "description": "Labels are key-value pairs for organizing tenants (optional for updates)",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "description": "Name is the updated tenant name (optional for updates)",
                    "type": "string"
                }
            }
        },
        "models.ComputeConfigDiscoveryResponse": {
            "type": "object",
            "properties": {
                "provider": {
                    "type": "string",
                    "description": "Provider is the active compute provider identifier (e.g., \"docker\")."
                },
                "schema": {
                    "type": "object",
                    "description": "Schema is the JSON Schema (draft 2020-12) for compute_config.",
                    "additionalProperties": true
                },
                "defaults": {
                    "type": "object",
                    "description": "Defaults is an optional defaults object for compute_config.",
                    "additionalProperties": true
                }
            }
        }
    }
}